Add and test handling of CCL integration errors
