+++ firecrown/connector/cosmosis/likelihood.py
@@ -143,7 +143,7 @@ class FirecrownLikelihood:
             # Cosmology will be generated by CCL. We want to be sure that CosmoSIS
             # has not also been configured to use CAMB.
             self.tools.prepare()
-        loglike = self.likelihood.compute_loglike(self.tools)
+        loglike = self.likelihood.compute_loglike_for_sampling(self.tools)
 
         derived_params_collection = self.likelihood.get_derived_parameters()
         assert derived_params_collection is not None
